+++ src/components/ui-data-list.vue
@@ -1,0 +1,5 @@
+<template>
+    <dl class="md:flex-wrap w-full md:flex border-t border-light-2">
+        <slot />
+    </dl>
+</template>