+++ src/components/ui-tabs.vue
@@ -1,0 +1,5 @@
+<template>
+    <div role="tablist" class="bg-white flex flex-wrap">
+        <slot />
+    </div>
+</template>